fix interderministic tests
