+++ litestar/_signature/models/attrs_signature_model.py
@@ -39,7 +39,7 @@ except ImportError as e:
 
 if TYPE_CHECKING:
     from litestar.plugins import PluginMapping
-    from litestar.types.parsed_signature import ParsedSignature
+    from litestar.utils.signature import ParsedSignature
 
 key_re = re.compile("@ attribute (.*)|'(.*)'")
