+++ litestar/utils/signature.py
@@ -1,27 +1,71 @@
 from __future__ import annotations
 
-from collections.abc import Collection
+import sys
+import typing
 from dataclasses import dataclass
-from inspect import Parameter, Signature
-from typing import TYPE_CHECKING, Any, AnyStr, Union
+from inspect import Parameter, Signature, isclass, ismethod
+from typing import Any, AnyStr, Collection, Union
 
-from typing_extensions import Annotated, NotRequired, Required, get_args, get_origin
+from typing_extensions import Annotated, NotRequired, Required, get_args, get_origin, get_type_hints
 
-from litestar.datastructures.state import ImmutableState
+from litestar.connection import Request, WebSocket
+from litestar.datastructures import Headers, ImmutableState, State
 from litestar.exceptions import ImproperlyConfiguredException
+from litestar.types import AnyCallable, Empty, Receive, Scope, Send, WebSocketScope
 from litestar.types.builtin_types import UNION_TYPES, NoneType
-from litestar.types.empty import Empty
-from litestar.utils.signature_parsing import get_fn_type_hints
 from litestar.utils.typing import get_safe_generic_origin, unwrap_annotation
 
-if TYPE_CHECKING:
-    from litestar.types import AnyCallable
+_GLOBAL_NAMES = {
+    "Headers": Headers,
+    "ImmutableState": ImmutableState,
+    "Receive": Receive,
+    "Request": Request,
+    "Scope": Scope,
+    "Send": Send,
+    "State": State,
+    "WebSocket": WebSocket,
+    "WebSocketScope": WebSocketScope,
+}
+"""A mapping of names used for handler signature forward-ref resolution.
 
-__all__ = (
-    "ParsedType",
-    "ParsedParameter",
-    "ParsedSignature",
-)
+This allows users to include these names within an `if TYPE_CHECKING:` block in their handler module.
+"""
+
+__all__ = ("get_fn_type_hints", "ParsedType", "ParsedParameter", "ParsedSignature")
+
+
+def get_fn_type_hints(fn: Any, namespace: dict[str, Any] | None = None) -> dict[str, Any]:
+    """Resolve type hints for ``fn``.
+
+    Args:
+        fn: Thing that is having its signature modelled.
+        namespace: Extra names for resolution of forward references.
+
+    Returns:
+        Mapping of names to types.
+    """
+    fn_to_inspect: Any = fn
+
+    if isclass(fn_to_inspect):
+        fn_to_inspect = fn_to_inspect.__init__
+
+    # detect objects that are not functions and that have a `__call__` method
+    if callable(fn_to_inspect) and ismethod(fn_to_inspect.__call__):
+        fn_to_inspect = fn_to_inspect.__call__
+
+    # inspect the underlying function for methods
+    if hasattr(fn_to_inspect, "__func__"):
+        fn_to_inspect = fn_to_inspect.__func__
+
+    # Order important. If a litestar name has been overridden in the function module, we want
+    # to use that instead of the litestar one.
+    namespace = {
+        **_GLOBAL_NAMES,
+        **vars(typing),
+        **vars(sys.modules[fn_to_inspect.__module__]),
+        **(namespace or {}),
+    }
+    return get_type_hints(fn_to_inspect, globalns=namespace, include_extras=True)
 
 
 @dataclass(frozen=True)
