+++ tests/kwargs/test_dto_extractor.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock
 
 from litestar._kwargs.extractors import create_dto_extractor
 from litestar.types.empty import Empty
-from litestar.types.parsed_signature import ParsedParameter, ParsedType
+from litestar.utils.signature import ParsedParameter, ParsedType
 from tests.dto import MockDTO, Model
